+++ charts/core/src/Chart/hooks/useChart.types.ts
@@ -19,12 +19,16 @@ export type ZoomSelect =
   | undefined;
 
 export interface ChartHookProps {
-  theme: Theme;
-
   /**
    * The id of the chart.
    */
   chartId?: string;
+
+  /**
+   * Enables tooltip synchronization across charts with the same `groupId`.
+   * `groupId` must be set for this to work.
+   */
+  enableGroupTooltipSync: boolean;
 
   /**
    * Charts with the same `groupId` will have their tooltips synced across charts.
@@ -37,11 +41,6 @@ export interface ChartHookProps {
   onChartReady?: () => void;
 
   /**
-   * Zoom selection enablement configuration.
-   */
-  zoomSelect?: ZoomSelect;
-
-  /**
    * Callback to be called when a zoom selection is made.
    */
   onZoomSelect?: (e: EChartZoomSelectionEvent) => void;
@@ -50,11 +49,18 @@ export interface ChartHookProps {
    * Controls the current chart state.
    */
   state?: ChartStates;
+
+  theme: Theme;
+
+  /**
+   * Zoom selection enablement configuration.
+   */
+  zoomSelect?: ZoomSelect;
 }
 
 export interface ChartInstance
   extends EChartsInstance,
-    Pick<ChartHookProps, 'state'>,
+    Pick<ChartHookProps, 'enableGroupTooltipSync' | 'state'>,
     Pick<
       UseTooltipVisibilityReturnObj,
       'isChartHovered' | 'setTooltipMounted' | 'tooltipPinned'