+++ packages/context-drawer/src/ButtonCorner/ButtonCorner.types.ts
@@ -1,0 +1,5 @@
+export const Side = {
+  Left: 'left',
+  Right: 'right',
+} as const;
+export type Side = (typeof Side)[keyof typeof Side];