+++ packages/context-drawer/src/ButtonCorner/index.ts
@@ -1,0 +1,2 @@
+export { ButtonCorner } from './ButtonCorner';
+export { Side } from './ButtonCorner.types';