+++ packages/tooltip/src/Tooltip/tooltipConstants.ts
@@ -3,4 +3,6 @@ import { transitionDuration } from '@leafygreen-ui/tokens';
 export const notchHeight = 8;
 export const notchWidth = 26;
 export const borderRadius = 16;
-export const hoverDelay = transitionDuration.slowest;
+
+export const DEFAULT_HOVER_DELAY = transitionDuration.slowest;
+export const CALLBACK_DEBOUNCE = 35; // ms