+++ tools/create/src/templates/component/src/testing/getTestUtils.index.template.ts
@@ -1,0 +1,4 @@
+export const testingIndex =
+  () => `export { getTestUtils } from './getTestUtils';
+export { type TestUtilsReturnType } from './getTestUtils.types';
+`;