+++ packages/descendants/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @leafygreen-ui/descendants
 
+## 3.0.1
+
+### Patch Changes
+
+- f685e51: Exports missing `UseHighlightOptions` and `HighlightContextType` types
+- Updated dependencies [164b15f]
+- Updated dependencies [518ce41]
+- Updated dependencies [3bef1e7]
+- Updated dependencies [3bef1e7]
+  - @leafygreen-ui/lib@15.1.0
+  - @leafygreen-ui/hooks@9.1.0
+  - @leafygreen-ui/leafygreen-provider@5.0.1
+
 ## 3.0.0
 
 ### Major Changes