+++ packages/drawer/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @leafygreen-ui/drawer
 
+## 4.0.1
+
+### Patch Changes
+
+- 164b15f: Use `LgIdString` type and remove setting `DEFAULT_LGID_ROOT`.
+- Updated dependencies [164b15f]
+- Updated dependencies [518ce41]
+- Updated dependencies [3bef1e7]
+- Updated dependencies [3bef1e7]
+- Updated dependencies [164b15f]
+- Updated dependencies [164b15f]
+- Updated dependencies [1eafbb5]
+- Updated dependencies [da277d5]
+  - @leafygreen-ui/lib@15.1.0
+  - @leafygreen-ui/hooks@9.1.0
+  - @leafygreen-ui/button@24.0.2
+  - @leafygreen-ui/typography@22.0.0
+  - @leafygreen-ui/tabs@16.0.0
+  - @leafygreen-ui/icon@14.1.0
+  - @lg-tools/test-harnesses@0.3.2
+  - @leafygreen-ui/icon-button@17.0.2
+  - @leafygreen-ui/leafygreen-provider@5.0.1
+  - @leafygreen-ui/palette@5.0.0
+  - @leafygreen-ui/polymorphic@3.0.2
+  - @leafygreen-ui/tokens@3.1.1
+  - @leafygreen-ui/toolbar@1.0.1
+
 ## 4.0.0
 
 ### Major Changes