+++ packages/stepper/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @leafygreen-ui/stepper
 
+## 6.0.2
+
+### Patch Changes
+
+- Updated dependencies [164b15f]
+- Updated dependencies [518ce41]
+- Updated dependencies [3bef1e7]
+- Updated dependencies [164b15f]
+- Updated dependencies [1eafbb5]
+- Updated dependencies [92c38b1]
+- Updated dependencies [92c38b1]
+  - @leafygreen-ui/lib@15.1.0
+  - @leafygreen-ui/typography@22.0.0
+  - @leafygreen-ui/icon@14.1.0
+  - @leafygreen-ui/tooltip@14.1.0
+  - @leafygreen-ui/leafygreen-provider@5.0.1
+  - @leafygreen-ui/palette@5.0.0
+  - @leafygreen-ui/tokens@3.1.1
+
 ## 6.0.1
 
 ### Patch Changes